requirements.txt
google-adk 
pandas 
requests
litellm